backend/pyproject.toml
[tool.poetry]
name = "job-portal"
version = "0.1.0"
description = ""
authors = ["smwasifzamanomee <wasifbdjsr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-main = "poetry.core.masonry.api"
